requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
blinker==1.7.0
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.3
Deprecated==1.2.14
exceptiongroup==1.2.0
fastapi==0.109.0
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.4.0
google-api-core==2.16.0
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.17
langsmith==0.0.84
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.20.2
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.16.3
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
posthog==3.3.4
proto-plus==1.23.0
protobuf==4.25.2
pulsar-client==3.4.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.0
pydantic_core==2.16.1
pydeck==0.8.1b0
Pygments==2.17.2
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-rapidjson==1.14
pytz==2023.4
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
streamlit==1.30.0
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.0
tornado==6.4
tqdm==4.66.1
transformers==4.37.2
tritonclient==2.42.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
uvicorn==0.27.0.post1
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
